@@ -1,20 +1,25 @@
 module.exports = {
   env: {
     browser: true,
+    commonjs: true,
     es2021: true,
   },
-  extends: [
-    "plugin:react/recommended",
-    "standard-with-typescript",
-    "next/core-web-vitals",
-    "prettier",
-  ],
+  extends: ['eslint:recommended', 'plugin:react/recommended', 'plugin:@typescript-eslint/recommended', 'plugin:prettier/recommended'],
   overrides: [],
+  parser: '@typescript-eslint/parser',
   parserOptions: {
-    ecmaVersion: "latest",
-    sourceType: "module",
-    project: ["tsconfig.json"],
+    ecmaFeatures: {
+      tsx: true,
+      jsx: true,
+    },
+    ecmaVersion: 'latest',
+    sourceType: 'module',
   },
-  plugins: ["react"],
-  rules: {},
+  plugins: ['react', '@typescript-eslint'],
+  rules: { 'react/react-in-jsx-scope': 'off', 'react/prop-types': 'off' },
+  settings: {
+    react: {
+      version: 'detect',
+    },
+  },
 };
