@@ -1,17 +1,11 @@
-import { NextApiRequest, NextApiResponse } from "next";
+import { NextApiRequest, NextApiResponse } from 'next';
 
 export interface ResponseType {
-  [key: string]: any;
+  [key: string]: unknown;
 }
 
-export default function withHandler(
-  method: "GET" | "POST",
-  fn: (req: NextApiRequest, res: NextApiResponse) => void
-) {
-  return async function (
-    req: NextApiRequest,
-    res: NextApiResponse
-  ): Promise<any> {
+export default function withHandler(method: 'GET' | 'POST', fn: (req: NextApiRequest, res: NextApiResponse) => void) {
+  return async function (req: NextApiRequest, res: NextApiResponse): Promise<unknown> {
     if (req.method !== method) {
       return res.status(405).end();
     }
