@@ -1,7 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  experimental: {
+    appDir: true,
+  },
   reactStrictMode: true,
-  output: "standalone",
+  output: 'standalone',
   async rewrites() {
     return [];
   },
